Refactor: Move DataFilter and FilterEntry into inventory package
